+++ src/main/java/insane96mcp/enhancedai/modules/mobs/ItemDisruption.java
@@ -1,5 +1,6 @@
-package insane96mcp.enhancedai.modules.enderman;
+package insane96mcp.enhancedai.modules.mobs;
 
+import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
@@ -7,23 +8,31 @@ import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.Difficulty;
 import insane96mcp.insanelib.base.config.LoadFeature;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
+import net.minecraft.tags.TagKey;
 import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.item.ItemEntity;
-import net.minecraft.world.entity.monster.EnderMan;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.phys.Vec2;
+import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.event.entity.living.LivingDamageEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-@Label(name = "Item Disruption", description = "Endermen will make the player's item fall from his hands.")
-@LoadFeature(module = Modules.Ids.ENDERMAN, enabledByDefault = false)
+@Label(name = "Item Disruption", description = "Endermen will make the player's item fall from his hands. Add/remove mobs via the enhancedai:can_disrupt_item entity type tag")
+@LoadFeature(module = Modules.Ids.MOBS)
 public class ItemDisruption extends Feature {
+    public static final TagKey<EntityType<?>> CAN_DISRUPT_ITEM = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "can_disrupt_item"));
+    public static final String ITEM_DISRUPTION_CHANCE = EnhancedAI.RESOURCE_PREFIX + "item_disruption_chance";
 
     @Config(min = 0d, max = 1d)
-    @Label(name = "Chance")
+    @Label(name = "Chance", description = "Chance can be changed within entity data's ForgeData.\"enhancedai:item_disruption_chance\"")
     public static Difficulty chance = new Difficulty(0.25d, 0.25d, 0.35d);
 
     public ItemDisruption(Module module, boolean enabledByDefault, boolean canBeDisabled) {
@@ -34,10 +43,10 @@ public class ItemDisruption extends Feature {
     public void onHit(LivingDamageEvent event) {
         if (!this.isEnabled()
                 || !(event.getEntity() instanceof ServerPlayer player)
-                || !(event.getSource().getDirectEntity() instanceof EnderMan enderman))
+                || !(event.getSource().getDirectEntity() instanceof Mob mob))
             return;
 
-        if (enderman.getRandom().nextFloat() < chance.getByDifficulty(player.level())) {
+        if (mob.getRandom().nextFloat() < mob.getPersistentData().getFloat(ITEM_DISRUPTION_CHANCE)) {
             ItemStack stack;
             ItemStack mainHandItem = player.getItemInHand(InteractionHand.MAIN_HAND);
             ItemStack offHandItem = player.getItemInHand(InteractionHand.OFF_HAND);
@@ -46,7 +55,7 @@ public class ItemDisruption extends Feature {
 
             if (!mainHandItem.isEmpty()) {
                 if (!offHandItem.isEmpty())
-                    stack = enderman.getRandom().nextBoolean() ? mainHandItem.copy() : offHandItem.copy();
+                    stack = mob.getRandom().nextBoolean() ? mainHandItem.copy() : offHandItem.copy();
                 else
                     stack = mainHandItem.copy();
             }
@@ -67,8 +76,27 @@ public class ItemDisruption extends Feature {
                 player.containerMenu.sendAllDataToRemote();
             });
             ItemEntity itementity = new ItemEntity(player.level(), player.getX(), player.getY() + player.getBbHeight() / 2f, player.getZ(), stack);
+            double x = player.getX() - mob.getX();
+            double z = player.getZ() - mob.getZ();
+            Vec2 dir = new Vec2((float) x, (float) z).normalized();
+            if (mob.getRandom().nextBoolean())
+                itementity.setDeltaMovement(itementity.getDeltaMovement().add(-dir.y * 0.4f, 0, dir.x * 0.4f));
+            else
+                itementity.setDeltaMovement(itementity.getDeltaMovement().add(dir.y * 0.4f, 0, -dir.x * 0.4f));
             itementity.setPickUpDelay(40);
             player.getCommandSenderWorld().addFreshEntity(itementity);
         }
+    }
+
+    @SubscribeEvent
+    public void onJoinLevelEvent(EntityJoinLevelEvent event) {
+        if (!this.isEnabled()
+                || event.getLevel().isClientSide
+                || !(event.getEntity() instanceof Mob mob)
+                || !mob.getType().is(CAN_DISRUPT_ITEM)
+                || mob.getPersistentData().contains(ITEM_DISRUPTION_CHANCE))
+            return;
+
+        mob.getPersistentData().putFloat(ITEM_DISRUPTION_CHANCE, (float) chance.getByDifficulty(mob.level()));
     }
 }